requirements.txt
numpy==1.20.1
pandas==1.4.1
mysql-connector-python==8.0.29
sqlalchemy==1.4.32